refactor(wire): the catalog rows and the txn frames are each defined once
Two wire layouts were spelled out on both sides of the connection.

**Catalog rows.** Every system-table row had two writers — one in the
client, one in the engine — and both were positional while both readers
were name-resolved against the shared `*_PAY_*` constants. Reordering a
family's column list therefore kept every reader correct and silently
transposed both writers, in two crates. gnitz_wire::sys_rows now owns one
writer per family (COL_TAB, TABLE_TAB, VIEW_TAB, IDX_TAB), taking a
named-field row struct and writing through a 2-method sink both sides
already had. A compile-time assert per family ties the emit order to the
payload constants, so a reordered column list fails the build instead. The
four "single home for this layout" comments that were false the moment the
second writer existed are gone with them.

**Txn frames.** PUSH_TXN, DDL_TXN and SCAN_MULTI were the last frame
families defined on one side only — matching code plus two near-verbatim
doc comments, with no shared constant pinning the u32 count, the mode
byte's position, the 10-byte relation record, or the 4 + 16n precondition
section. gnitz_wire::txn_frame owns both directions; the client keeps its
public wrappers as adapters that pre-encode each block, and the engine
re-exports the decoders rather than wrapping them.

Encoded bytes are unchanged. New unit tests in gnitz-wire cover the frames
that had none of their own: a truncation sweep at every offset of all three
frames, hostile item and precondition counts, and the control-header
properties (only the routing flag, zero seek fields) the client's
hand-written frame walk used to assert on its way past.

The client's pack_col_id wrapper and its three tests moved down to the
codec they exercise; `IndexRecord.is_unique` stays the stored word rather
than a bool, since a DROP must echo back exactly what it read.

+++ crates/gnitz-engine/src/runtime/tests/wire.rs
@@ -1182,8 +1182,8 @@ fn decode_ddl_txn_is_unaffected_by_the_appended_section() {
     let payload = gnitz_core::protocol::encode_ddl_txn(0xABCD, &[(SCHEMA_TAB, sch_b), (SEQ_TAB, seq_b)]);
     let decoded = decode_ddl_txn(&payload).unwrap();
     assert_eq!(decoded.len(), 2);
-    assert_eq!(decoded[0].0, SCHEMA_TAB as i64);
-    assert_eq!(decoded[1].0, SEQ_TAB as i64);
+    assert_eq!(decoded[0].0, SCHEMA_TAB as u32);
+    assert_eq!(decoded[1].0, SEQ_TAB as u32);
 }
 
 #[test]
